perf(prometheus): streamline unified parser driver
